Aggiungi AssignmentService per la dashboard consegne
